requirements.txt
google-cloud-translate==0.22.0
git+https://github.com/joestump/python-oauth2.git
redis
